+++ assets/js/theme/product/click-and-buy/App.js
@@ -21,6 +21,7 @@ import MarketingBtn from '../react-components/MarketingBtn';
 import PointOfPurchaseModal from '../react-components/Modal-PointOfPurchase';
 
 
+
 class App extends React.Component{
     constructor(props){
         super()